separe typedefs and schema in graphql
